Upgraded projects to dotnet core 3.0. This was done since 3.0 was needed inorder to handle new logic with certificates.
On Openshift, we couldn't get the old proven way with using p12 (or pfx) to work. So now the apporach is to instead work with the cert and (unencrypted) private key in pem format and jump through some hoops in the code to make .net core swallow that way of loading the certificate. Not sure if it work yet though...